// react-ui/types/generated/models/road-model.ts
/**
 * This is a TypeGen auto-generated file.
 * Any changes made to this file can be lost when this file is regenerated.
 */

import { RoadKey } from './road-key';
import { RoadState } from './road-state';

export interface RoadModel {
    roadKey: RoadKey;
    roadState: RoadState;
    ownerId: string;
    buildIndex: number;
}