// react-ui/types/generated/models/entitlement.ts
/**
 * This is a TypeGen auto-generated file.
 * Any changes made to this file can be lost when this file is regenerated.
 */

export type Entitlement = 'Undefined' | 'DevCard' | 'Settlement' | 'City' | 'Road' | 'Ship' | 'BuyKnight' | 'UpgradeKnight' | 'ActivateKnight' | 'Soldier' | 'PoliticsUpgrade' | 'ScienceUpgrade' | 'TradeUpgrade' | 'Wall' | 'DestroyCity' | 'Bishop' | 'Deserter' | 'Inventor' | 'Intrigue' | 'Diplomat' | 'Merchant' | 'KnightDisplacement' | 'UpgradeToMetro' | 'KnightDisplacementMoveKnightOutOfTheWay' | 'RolledSeven';

export const Entitlement = {
    Undefined: 'Undefined',
    DevCard: 'DevCard',
    Settlement: 'Settlement',
    City: 'City',
    Road: 'Road',
    Ship: 'Ship',
    BuyKnight: 'BuyKnight',
    UpgradeKnight: 'UpgradeKnight',
    ActivateKnight: 'ActivateKnight',
    Soldier: 'Soldier',
    PoliticsUpgrade: 'PoliticsUpgrade',
    ScienceUpgrade: 'ScienceUpgrade',
    TradeUpgrade: 'TradeUpgrade',
    Wall: 'Wall',
    DestroyCity: 'DestroyCity',
    Bishop: 'Bishop',
    Deserter: 'Deserter',
    Inventor: 'Inventor',
    Intrigue: 'Intrigue',
    Diplomat: 'Diplomat',
    Merchant: 'Merchant',
    KnightDisplacement: 'KnightDisplacement',
    UpgradeToMetro: 'UpgradeToMetro',
    KnightDisplacementMoveKnightOutOfTheWay: 'KnightDisplacementMoveKnightOutOfTheWay',
    RolledSeven: 'RolledSeven',
} as const;